coffee-machine/requirements.txt
# Job Automation Tool - Requirements
# Python 3.11+

# Core Framework
crewai>=0.28.0
crewai-tools>=0.2.0

# Web Frontend
streamlit>=1.30.0
streamlit-option-menu>=0.3.0

# AI/LLM
openai>=1.3.0

# Monitoring
agentops>=0.1.0

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0

# Document Export
python-docx>=0.8.11
reportlab>=4.0.0

# Data Visualization
plotly>=5.0.0
pandas>=2.0.0

# Additional Tools
beautifulsoup4>=4.12.0
requests>=2.31.0
python-kasa>=0.6.0
